refactor
handle missing asset
