+++ mobile/ios/Runner/Resolvers/Images/Thumbnails.g.swift
